refactor: refactoring list ♻️
