+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/Commands/Commands.module.scss
@@ -1,0 +1,23 @@
+@import "~@fluentui/react/dist/sass/References.scss";
+
+.commands {
+  display: flex;
+  flex-direction: row;
+  max-width: 1024px;
+  gap: 15px;
+
+  .search {
+    width: 100%;
+    max-width: 100%;
+
+    .searchBox {
+      box-shadow: var(--shadow2);
+      width: inherit;
+      max-width: inherit;
+    }
+  }
+
+  .sortBy {
+    box-shadow: var(--shadow2);
+  }
+}
