+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/Commands/index.ts
@@ -1,0 +1,1 @@
+export * from './Commands'
