+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/List/index.ts
@@ -1,0 +1,2 @@
+export * from './List'
+export * from './context'
