+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/List/types.ts
@@ -1,0 +1,8 @@
+import { TableColumnDefinition } from "@fluentui/react-components";
+import { ProjectListModel } from "pp365-shared-library";
+
+
+export interface IListColumn extends TableColumnDefinition<ProjectListModel> {
+  minWidth?: number;
+  defaultWidth?: number;
+}
