+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectMenu/index.ts
@@ -1,0 +1,2 @@
+export * from './ProjectMenu'
+export * from './types'
