+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/context.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+import { useProjectList } from './useProjectList';
+
+export interface IProjectListContext extends ReturnType<typeof useProjectList> {
+
+}
+
+export const ProjectListContext = createContext<IProjectListContext>(null)
